+++ NEWS.md
@@ -1,5 +1,7 @@
 # vip 0.1.2
 
+* Added `nsim` argument to `vi_permute()` for reducing the sampling variability induced by permuting each predictor [(#36)](https://github.com/koalaverse/vip/issues/36).
+
 * Greatly improved the documentation for `vi_model()` and the various objects it supports.
 
 * New argument `rank`, which default to `FALSE`, available in `vi()` [(#55)](https://github.com/koalaverse/vip/issues/55).