+++ NEWS.md
@@ -6,7 +6,7 @@
 
 * Greatly improved the documentation for `vi_model()` and the various objects it supports.
 
-* New argument `rank`, which default to `FALSE`, available in `vi()` [(#55)](https://github.com/koalaverse/vip/issues/55).
+* New argument `rank`, which defaults to `FALSE`, available in `vi()` [(#55)](https://github.com/koalaverse/vip/issues/55).
 
 * Added support for Spark (G)LMs.